completed ability to extract stackexchange q and a
